Fix harmless compiler warnings in testing code.
FossilOrigin-Name: 2f53612388507f9b1ecbec3c15281825beab54b8062d97d0ab6f009c5fed347b

+++ ext/recover/dbdata.c
@@ -697,9 +697,9 @@ static int dbdataEof(sqlite3_vtab_cursor *pCursor){
 ** otherwise.
 */
 static int dbdataIsFunction(const char *zSchema){
-  int n = strlen(zSchema);
+  size_t n = strlen(zSchema);
   if( n>2 && zSchema[n-2]=='(' && zSchema[n-1]==')' ){
-    return n-2;
+    return (int)n-2;
   }
   return 0;
 }
